requirements.txt
argparse
python-dotenv
google-generativeai